backend/pyproject.toml
[tool.poetry]
name = "template"
version = "0.1.0"
description = ""
authors = ["Your Name <youremail@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
fastapi = "^0.112.0"
uvicorn = "^0.29.0"
pydantic-settings = "^2.7.1"
apscheduler = "^3.11.0"
langchain-openai = "^0.3.0"
langchain = "^0.3.14"
langchain-core = "^0.3.29"
langgraph = "^0.2.62"
langchain-community = "^0.3.14"
APScheduler = "^3.11.0"
python-telegram-bot = "^21.10"
googlesearch-python = "^1.3.0"
dexscreener = "^1.2"
mem0ai = "^0.1.93"
timeout-decorator = "^0.5.0"
fake_useragent = "^1.5.1"
python-dotenv = "^1.0.0"
cachetools = "^5.3.3"
python-multipart = "^0.0.7"
PyMuPDF = "^1.24.4"
pdfplumber = "^0.10.3"
pypdf = "^4.2.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
